[JENKINS-36899] Set estimatedWorkspaceSize as allocation strategy parameter
