+++ src/main/resources/org/jenkinsci/plugins/ewm/strategies/FastestReadSpeedStrategy/config.jelly
@@ -1,0 +1,6 @@
+<?jelly escape-by-default='true'?>
+<j:jelly xmlns:j="jelly:core" xmlns:f="/lib/form">
+    <f:entry title="${%Estimated Workspace Size [KB]}" field="estimatedWorkspaceSize">
+        <f:number/>
+    </f:entry>
+</j:jelly>
